chore: Move theme further down render tree
